refactor the directory structure
