+++ clients/cli/src/config/mod.rs
@@ -1,0 +1,2 @@
+pub mod general;
+pub mod prover;
