+++ clients/cli/src/config/prover.rs
@@ -1,15 +1,26 @@
-// use crate::generated;
-
-use crate::prover_id_manager;
-
-use crate::analytics::track;
-use serde_json::json;
+use crate::utils::analytics::track;
+use crate::utils::prover_id::get_or_generate_prover_id;
 
 use nexus_core::prover::nova::{
     pp::gen_vm_pp,
     types::{seq, PublicParams, C1, C2, G1, G2, RO, SC},
 };
+use serde_json::json;
+// use serial_test::serial;
+// use std::sync::Once;
+use tracing_subscriber::fmt::format::FmtSpan;
+use tracing_subscriber::EnvFilter;
 
+/// Configuration for the RISC-V zero-knowledge prover
+///
+/// This struct holds essential parameters used throughout the proving process:
+/// - `prover_id`: Unique identifier for this prover instance
+/// - `k`: Step size for the proving system (number of cycles per step)
+/// - `ws_addr_string`: WebSocket address for connecting to the orchestrator
+/// - `public_parameters`: Nova-specific parameters for generating zero-knowledge proofs
+///
+/// Used in main.rs to initialize the prover and maintain connection settings
+/// throughout the proving lifecycle.
 pub struct ProverConfig {
     pub prover_id: String,
     pub k: i32,
@@ -22,6 +33,13 @@ pub async fn initialize(
     hostname: String,
     port: u16,
 ) -> Result<ProverConfig, Box<dyn std::error::Error>> {
+    // Configure the tracing subscriber
+    // This is a global value so we do not need to pass it around
+    tracing_subscriber::fmt()
+        .with_env_filter(EnvFilter::from_default_env())
+        .with_span_events(FmtSpan::CLOSE)
+        .init();
+
     // Construct the WebSocket URL based on the port number
     // Uses secure WebSocket (wss) for port 443, regular WebSocket (ws) otherwise
     let ws_addr_string = format!(
@@ -33,12 +51,14 @@ pub async fn initialize(
 
     // Set the constant k value used for proof generation
     // This determines the size/complexity of the proving system
+    // Higher values increase proof generation speed but require more memory
     let k = 4;
 
     // Retrieve an existing prover ID from storage or generate a new one
     // This ID uniquely identifies this prover instance
-    let prover_id = prover_id_manager::get_or_generate_prover_id();
+    let prover_id = get_or_generate_prover_id();
 
+    // Track the registration event
     track(
         "register".into(),
         format!("Your assigned prover identifier is {}.", prover_id),
@@ -60,14 +80,6 @@ pub async fn initialize(
         Err(e) => return Err(format!("Failed to generate public parameters: {}", e).into()),
     };
 
-    track(
-        "connect".into(),
-        format!("Connecting to {}...", &ws_addr_string),
-        &ws_addr_string,
-        json!({"prover_id": prover_id}),
-    );
-
-    // Construct and return the ProverConfig with the initialized values
     Ok(ProverConfig {
         ws_addr_string,
         k,
