+++ clients/cli/src/network/mod.rs
@@ -1,0 +1,2 @@
+pub mod connection;
+pub mod websocket;
