+++ clients/cli/src/network/websocket.rs
