+++ clients/cli/src/utils/prover_id.rs
